pyproject.toml
# ~~ Generated by projen. To modify, edit .projenrc.py and run "npx projen".

[tool.poetry]
name = "bedrock_hitl"
version = "0.1.0"
description = ""
authors = [ "ArJSarmiento <rneljan@gmail.com>" ]
readme = "README.md"

  [tool.poetry.dependencies]
  aws-cdk-lib = "^2.1.0"
  constructs = "^10.0.5"
  python = "^3.8"

[tool.poetry.group.dev.dependencies]
projen = "0.85.0"
pytest = "7.4.3"

[tool.poetry.group.format.dependencies]
ruff = "0.5.1"

[build-system]
requires = [ "poetry-core" ]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 120
target-version = "py310"

[tool.ruff.format]
quote-style = "single"
docstring-code-line-length = "dynamic"
indent-style = "space"
docstring-code-format = false
line-ending = "auto"


[tool.ruff.lint]
select = ["E4", "E7", "E9", "F"]
ignore = []